Fix cost to go calculation.
The rewire step was invalidating memoized children cost to go.
The paper implies that cost to go can be computed recursively without
memoization without affecting algorithmic complexity, so I did it that
way and it seems to be fine.

A separate function is provided which computes cost to go in one
sweep. This requires tracking children in addition to parents, so a
map of parent->children is added to Search.